Add maxConcurrency argument to NewDispatchingHandler to rate limit active goroutines
Previously, the number of goroutines created by calls to DispatchEvent()
was unrestricted. This commit adds a new argument to
NewDispatchingHandler, backed by a new dispatchLimiter struct, that
limits the number of goroutines that DispatchEvent() can have in flight
at one time.

This is based heavily off the design of LimitListener from golang/net

Fixes #24