This fixes TestCloudHandlerExpirationAndRefresh
The core of this fix is to effectively de-duplicate the list of IPs
refreshed, so it doesn't matter if we refresh 2 or 3 times.  However
as there was also a 900ms sleep, this also trims that down and adds a
comment explaining the theoretical timing breakdown.  It also serializes
the test to try and minimize other CPU load on the test.

The real fix would involve a virtual clock, but that's too much effort
for now.

Fixes #111