Refactor how coverage data is generated
I THINK we originally did it this way because earlier versions of go could
only cover one package at a time.  Either way, this works and allows the
coverage tests to run in parallel, taking us from ~23s to ~10s to run.

This also re-orders the build order to put things most likely to be fast,
both from a completion and a failure view, towards the top.