Remove MetricAggregator.Receive and BackendHandler.DispatchMetrics, with associated dead code cleanup.
This also removes a pair of benchmarks, one for hot metrics, the other for splitting metrics
up.  These are not relevant anymore.  Hot metrics will be handled at an earlier point before
there is contention, and splitting MetricMaps is not a high volume act. If it is benchmarked
it should be a MetricMap benchmark, not a BackendHandler benchmark.