Only run slow tests by hand
We currently have two slow tests, one is a benchmark in the form of a
test, and one is an actual benchmark.  ARM64 is having a bad time with
the benchmark, so we'll skip it for now until it can be fixed.  This
does *not* close the issue attached to it.