Fixes race conditions in the tests
SUBSCRIBE and PUBLISH have their own connections, which results in a race
for which one executes first.  This forces the SUBSCRIBE connection to
acknowledge the subscription before PUBLISH proceeds.  It also loosens up
the timings a bit, and removes a direct race condition where the context
is canceled in TestRedisNodeTrackerSelf.