Rewrite node tracking (#566)
* Rewrite node tracking

This should be considered as a from scratch rewrite of the node tracking stuff,
and not a change.  Nothing uses it, and the whole approach has changed.

* Fixes race conditions in the tests

SUBSCRIBE and PUBLISH have their own connections, which results in a race
for which one executes first.  This forces the SUBSCRIBE connection to
acknowledge the subscription before PUBLISH proceeds.  It also loosens up
the timings a bit, and removes a direct race condition where the context
is canceled in TestRedisNodeTrackerSelf.